Newstool/requirements.txt
python-dotenv==1.2.1
langchain-core==1.2.7
langchain-community== 0.4.1
langchain-text-splitters== 1.1.0
langchain-chroma== 1.1.0
langchain-groq== 1.1.1
chromadb== 1.4.1
huggingface-hub== 0.36.0
sentence-transformers== 2.6.1
